calendar fix inject error
