+++ app/src/main/AndroidManifest.xml
@@ -3,6 +3,7 @@
     package="com.cornellappdev.scoop">
 
     <application
+        android:name = ".ui.CalendarApplication"
         android:allowBackup="true"
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
