+++ app/src/main/java/com/cornellappdev/scoop/ui/CalendarApplication.kt
@@ -1,0 +1,8 @@
+package com.cornellappdev.scoop.ui
+
+import android.app.Application
+import dagger.hilt.android.HiltAndroidApp
+
+@HiltAndroidApp
+class CalendarApplication: Application() {
+}
