+++ app/src/main/java/com/cornellappdev/scoop/ui/MainActivity.kt
@@ -4,18 +4,98 @@ import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.compose.animation.ExperimentalAnimationApi
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.material.MaterialTheme
+import androidx.compose.material.Scaffold
+import androidx.compose.material.Text
+import androidx.compose.material.TopAppBar
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.text.style.TextAlign
+import androidx.compose.ui.unit.dp
+import androidx.lifecycle.viewmodel.compose.viewModel
+import com.cornellappdev.scoop.ui.components.calendar.*
+import com.cornellappdev.scoop.ui.components.general.Calendar
 import com.cornellappdev.scoop.ui.navigation.MainScreen
+import com.cornellappdev.scoop.ui.theme.ScoopTheme
+import com.google.accompanist.insets.ProvideWindowInsets
+import com.google.accompanist.insets.statusBarsHeight
 import com.google.accompanist.pager.ExperimentalPagerApi
+import dagger.hilt.android.AndroidEntryPoint
 
+@AndroidEntryPoint
 class MainActivity : ComponentActivity() {
-
-    @OptIn(
-        ExperimentalAnimationApi::class, ExperimentalPagerApi::class
-    )
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
+
         setContent {
-            MainScreen()
+            ProvideWindowInsets {
+                ScoopTheme {
+                    CalendarScreen(onBackPressed = { finish() })
+                }
+            }
         }
+    }
+}
+
+@Composable
+fun CalendarScreen(
+    onBackPressed: () -> Unit
+) {
+    val calendarViewModel: CalendarViewModel = viewModel()
+    val calendarYear = calendarViewModel.calendarYear
+
+    CalendarContent(
+        selectedDates = calendarViewModel.datesSelected.toString(),
+        calendarYear = calendarYear,
+        onDayClicked = { calendarDay, calendarMonth ->
+            calendarViewModel.onDaySelected(
+                DaySelected(calendarDay.value.toInt(), calendarMonth, calendarYear)
+            )
+        },
+        onBackPressed = onBackPressed
+    )
+}
+
+@Composable
+fun CalendarContent(
+    selectedDates: String,
+    calendarYear: CalendarYear,
+    onDayClicked: (CalendarDay, CalendarMonth) -> Unit,
+    onBackPressed: () -> Unit
+) {
+    Scaffold(
+        backgroundColor = Color.White,
+        topBar = {
+            CalendarTopAppBar(selectedDates, onBackPressed)
+        }
+    ) {
+        Calendar(calendarYear, onDayClicked)
+    }
+}
+
+@Composable
+fun CalendarTopAppBar(selectedDates: String, onBackPressed: () -> Unit) {
+    Column {
+        Spacer(modifier = Modifier
+            .statusBarsHeight()
+            .fillMaxWidth()
+            .background(MaterialTheme.colors.primaryVariant)
+        )
+        TopAppBar(
+            title = {
+                Text(
+                    text = "Departure Date",
+                    modifier = Modifier.fillMaxWidth(),
+                    textAlign = TextAlign.Center
+                )
+            },
+            backgroundColor = Color.White,
+            elevation = 0.dp
+        )
     }
 }
