+++ app/src/main/java/com/cornellappdev/scoop/ui/theme/Color.kt
@@ -7,6 +7,7 @@ val Gray = Color(0xFFE8E8E8)
 val DarkGray = Color(0xFFAFACAC)
 val LightGray = Color(0xFFECECEC)
 val PlaceholderGray = Color(0xFFC4C4C4)
+val ScoopGreen = Color(0x60BFA0)
 
 val ColorPalette = lightColors(
     background = Color.White,
